# src/main/java/com/app/booking/controller/CustomerController.java
/*
package com.app.booking.controller;

import com.app.booking.service.api.CustomerApi;
import com.app.booking.service.model.TrackCustomer;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.RestController;

import java.util.List;

@RestController
public class CustomerController implements CustomerApi {
    @Override
    public ResponseEntity<Object> executeTrigger(String fileName, String source) {
        return ResponseEntity.status(HttpStatus.NOT_IMPLEMENTED).build();
    }

    @Override
    public ResponseEntity<List<TrackCustomer>> fetchCustomer(String propertyName) {
        return ResponseEntity.status(HttpStatus.NOT_IMPLEMENTED).build();
    }
}
*/